+++ src/libraries/System.Private.CoreLib/src/System/Threading/Interlocked.cs
@@ -1,0 +1,259 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.CompilerServices;
+using Internal.Runtime.CompilerServices;
+
+namespace System.Threading
+{
+    /// <summary>Provides atomic operations for variables that are shared by multiple threads.</summary>
+    public static partial class Interlocked
+    {
+        #region Increment
+        /// <summary>Increments a specified variable and stores the result, as an atomic operation.</summary>
+        /// <param name="location">The variable whose value is to be incremented.</param>
+        /// <returns>The incremented value.</returns>
+        /// <exception cref="NullReferenceException">The address of location is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint Increment(ref uint location) =>
+            Add(ref location, 1);
+
+        /// <summary>Increments a specified variable and stores the result, as an atomic operation.</summary>
+        /// <param name="location">The variable whose value is to be incremented.</param>
+        /// <returns>The incremented value.</returns>
+        /// <exception cref="NullReferenceException">The address of location is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Increment(ref ulong location) =>
+            Add(ref location, 1);
+        #endregion
+
+        #region Decrement
+        /// <summary>Decrements a specified variable and stores the result, as an atomic operation.</summary>
+        /// <param name="location">The variable whose value is to be decremented.</param>
+        /// <returns>The decremented value.</returns>
+        /// <exception cref="NullReferenceException">The address of location is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint Decrement(ref uint location) =>
+            (uint)Add(ref Unsafe.As<uint, int>(ref location), -1);
+
+        /// <summary>Decrements a specified variable and stores the result, as an atomic operation.</summary>
+        /// <param name="location">The variable whose value is to be decremented.</param>
+        /// <returns>The decremented value.</returns>
+        /// <exception cref="NullReferenceException">The address of location is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Decrement(ref ulong location) =>
+            (ulong)Add(ref Unsafe.As<ulong, long>(ref location), -1);
+        #endregion
+
+        #region Exchange
+        /// <summary>Sets a 32-bit unsigned integer to a specified value and returns the original value, as an atomic operation.</summary>
+        /// <param name="location1">The variable to set to the specified value.</param>
+        /// <param name="value">The value to which the <paramref name="location1"/> parameter is set.</param>
+        /// <returns>The original value of <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of location1 is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint Exchange(ref uint location1, uint value) =>
+            (uint)Exchange(ref Unsafe.As<uint, int>(ref location1), (int)value);
+
+        /// <summary>Sets a 64-bit unsigned integer to a specified value and returns the original value, as an atomic operation.</summary>
+        /// <param name="location1">The variable to set to the specified value.</param>
+        /// <param name="value">The value to which the <paramref name="location1"/> parameter is set.</param>
+        /// <returns>The original value of <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of location1 is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Exchange(ref ulong location1, ulong value) =>
+            (ulong)Exchange(ref Unsafe.As<ulong, long>(ref location1), (long)value);
+        #endregion
+
+        #region CompareExchange
+        /// <summary>Compares two 32-bit unsigned integers for equality and, if they are equal, replaces the first value.</summary>
+        /// <param name="location1">The destination, whose value is compared with <paramref name="comparand"/> and possibly replaced.</param>
+        /// <param name="value">The value that replaces the destination value if the comparison results in equality.</param>
+        /// <param name="comparand">The value that is compared to the value at <paramref name="location1"/>.</param>
+        /// <returns>The original value in <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint CompareExchange(ref uint location1, uint value, uint comparand) =>
+            (uint)CompareExchange(ref Unsafe.As<uint, int>(ref location1), (int)value, (int)comparand);
+
+        /// <summary>Compares two 64-bit unsigned integers for equality and, if they are equal, replaces the first value.</summary>
+        /// <param name="location1">The destination, whose value is compared with <paramref name="comparand"/> and possibly replaced.</param>
+        /// <param name="value">The value that replaces the destination value if the comparison results in equality.</param>
+        /// <param name="comparand">The value that is compared to the value at <paramref name="location1"/>.</param>
+        /// <returns>The original value in <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong CompareExchange(ref ulong location1, ulong value, ulong comparand) =>
+            (ulong)CompareExchange(ref Unsafe.As<ulong, long>(ref location1), (long)value, (long)comparand);
+        #endregion
+
+        #region Add
+        /// <summary>Adds two 32-bit unsigned integers and replaces the first integer with the sum, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be added. The sum of the two values is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be added to the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint Add(ref uint location1, uint value) =>
+            (uint)Add(ref Unsafe.As<uint, int>(ref location1), (int)value);
+
+        /// <summary>Adds two 64-bit unsigned integers and replaces the first integer with the sum, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be added. The sum of the two values is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be added to the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Add(ref ulong location1, ulong value) =>
+            (ulong)Add(ref Unsafe.As<ulong, long>(ref location1), (long)value);
+        #endregion
+
+        #region Read
+        /// <summary>Returns a 64-bit unsigned value, loaded as an atomic operation.</summary>
+        /// <param name="location">The 64-bit value to be loaded.</param>
+        /// <returns>The loaded value.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Read(ref ulong location) =>
+            CompareExchange(ref location, 0, 0);
+        #endregion
+
+        #region And
+        /// <summary>Bitwise "ands" two 32-bit signed integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int And(ref int location1, int value)
+        {
+            int current = location1;
+            while (true)
+            {
+                int newValue = current & value;
+                int oldValue = CompareExchange(ref location1, newValue, current);
+                if (oldValue == current)
+                {
+                    return newValue;
+                }
+                current = oldValue;
+            }
+        }
+
+        /// <summary>Bitwise "ands" two 32-bit unsigned integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint And(ref uint location1, uint value) =>
+            (uint)And(ref Unsafe.As<uint, int>(ref location1), (int)value);
+
+        /// <summary>Bitwise "ands" two 64-bit signed integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static long And(ref long location1, long value)
+        {
+            long current = location1;
+            while (true)
+            {
+                long newValue = current & value;
+                long oldValue = CompareExchange(ref location1, newValue, current);
+                if (oldValue == current)
+                {
+                    return newValue;
+                }
+                current = oldValue;
+            }
+        }
+
+        /// <summary>Bitwise "ands" two 64-bit unsigned integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong And(ref ulong location1, ulong value) =>
+            (ulong)And(ref Unsafe.As<ulong, long>(ref location1), (long)value);
+        #endregion
+
+        #region Or
+        /// <summary>Bitwise "ors" two 32-bit signed integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static int Or(ref int location1, int value)
+        {
+            int current = location1;
+            while (true)
+            {
+                int newValue = current | value;
+                int oldValue = CompareExchange(ref location1, newValue, current);
+                if (oldValue == current)
+                {
+                    return newValue;
+                }
+                current = oldValue;
+            }
+        }
+
+        /// <summary>Bitwise "ors" two 32-bit unsigned integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static uint Or(ref uint location1, uint value) =>
+            (uint)Or(ref Unsafe.As<uint, int>(ref location1), (int)value);
+
+        /// <summary>Bitwise "ors" two 64-bit signed integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static long Or(ref long location1, long value)
+        {
+            long current = location1;
+            while (true)
+            {
+                long newValue = current | value;
+                long oldValue = CompareExchange(ref location1, newValue, current);
+                if (oldValue == current)
+                {
+                    return newValue;
+                }
+                current = oldValue;
+            }
+        }
+
+        /// <summary>Bitwise "ors" two 64-bit unsigned integers and replaces the first integer with the result, as an atomic operation.</summary>
+        /// <param name="location1">A variable containing the first value to be combined. The result is stored in <paramref name="location1"/>.</param>
+        /// <param name="value">The value to be combined with the integer at <paramref name="location1"/>.</param>
+        /// <returns>The new value stored at <paramref name="location1"/>.</returns>
+        /// <exception cref="NullReferenceException">The address of <paramref name="location1"/> is a null pointer.</exception>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [CLSCompliant(false)]
+        public static ulong Or(ref ulong location1, ulong value) =>
+            (ulong)Or(ref Unsafe.As<ulong, long>(ref location1), (long)value);
+        #endregion
+    }
+}
